float

+++ opennn/config.h
@@ -36,6 +36,6 @@
 
 namespace OpenNN
 {
-    typedef double type;
+    typedef float type;
 }
